Revert 32-bit wrap-around fix and solve the edge case that triggered it instead, re issue #232

+++ src/internal.h
@@ -512,7 +512,7 @@ struct trail_ {
 
 struct slot_ {
 	cell c;
-	uint64_t vgen, vgen2;	// visit generation
+	uint32_t vgen, vgen2;	// visit generation
 };
 
 // Where 'prev_offset' is the number of frames back
@@ -671,7 +671,7 @@ struct query_ {
 	pl_idx_t q_size[MAX_QUEUES], tmpq_size[MAX_QUEUES], qp[MAX_QUEUES];
 	prolog_flags flags;
 	enum q_retry retry;
-	uint64_t vgen;
+	uint32_t vgen;
 	int8_t halt_code;
 	int8_t quoted;
 	bool done:1;
